+++ R/utilities.R
@@ -1,16 +1,4 @@
-## TODO: Add comment
-#
-## Author: David Carslaw
-## useful utility functions
-## with some updates and modification by Karl Ropkins
-###############################################################################
-
-startYear <- function(dat) as.numeric(format(min(dat[order(dat)]), "%Y"))
-endYear <- function(dat) as.numeric(format(max(dat[order(dat)]), "%Y"))
-startMonth <- function(dat) as.numeric(format(min(dat[order(dat)]), "%m"))
-endMonth <- function(dat) as.numeric(format(max(dat[order(dat)]), "%m"))
-
-## these are pre-defined type that need a field "date"; used by cutData
+# these are pre-defined type that need a field "date"; used by cutData
 dateTypes <- c(
   "year",
   "hour",
@@ -28,19 +16,19 @@ dateTypes <- c(
   "yearseason"
 )
 
-## sets up how openair graphics look by default and resets on exit
-
-setGraphics <- function(fontsize = 5) {
-  current.strip <- trellis.par.get("strip.background")
-  trellis.par.set(fontsize = list(text = fontsize))
-
-  ## reset graphic parameters
-  font.orig <- trellis.par.get("fontsize")$text
-  on.exit(trellis.par.set(
-    fontsize = list(text = font.orig)
-  ))
+# get date components
+startYear <- function(dat) {
+  as.numeric(format(min(dat[order(dat)]), "%Y"))
 }
-
+endYear <- function(dat) {
+  as.numeric(format(max(dat[order(dat)]), "%Y"))
+}
+startMonth <- function(dat) {
+  as.numeric(format(min(dat[order(dat)]), "%m"))
+}
+endMonth <- function(dat) {
+  as.numeric(format(max(dat[order(dat)]), "%m"))
+}
 
 # function to test of a suggested package is available and warn if not
 try_require <- function(package, fun) {
@@ -60,17 +48,14 @@ try_require <- function(package, fun) {
   )
 }
 
-###############################################################################
-
-## function to find averaging period of data, returns "xx sec"
-## for use in filling in gaps in time series data
-## it finds the table of values of time gaps and picks the biggest
-## can't think of better way unless user specifies what the time interval is meant to be
-
+# function to find averaging period of data, returns "xx sec"
+# for use in filling in gaps in time series data
+# it finds the table of values of time gaps and picks the biggest
+# can't think of better way unless user specifies what the time interval is meant to be
 find.time.interval <- function(dates) {
-  ## could have several sites, dates may be unordered
-  ## find the most common time gap in all the data
-  dates <- unique(dates) ## make sure they are unique
+  # could have several sites, dates may be unordered
+  # find the most common time gap in all the data
+  dates <- unique(dates) # make sure they are unique
 
   # work out the most common time gap of unique, ordered dates
   id <- which.max(table(diff(as.numeric(unique(dates[order(dates)])))))
@@ -88,48 +73,46 @@ find.time.interval <- function(dates) {
   seconds
 }
 
-## #################################################################
 # Function to pad out missing time data
 # assumes data have already been split by type, so just take first
 # tries to work out time interval of input based on most common gap
 # can print assumed gap to screen
-
 date.pad <- function(mydata, type = NULL, print.int = FALSE) {
   # if one line, just return
   if (nrow(mydata) < 2) {
     return(mydata)
   }
 
-  ## time zone of data
+  # time zone of data
   TZ <- attr(mydata$date, "tzone")
   if (is.null(TZ)) {
     TZ <- "GMT"
-  } ## as it is on Windows for BST
+  } # as it is on Windows for BST
 
-  ## function to fill missing data gaps
-  ## assume no missing data to begin with
+  # function to fill missing data gaps
+  # assume no missing data to begin with
 
-  ## pad out missing data
+  # pad out missing data
   start.date <- min(mydata$date, na.rm = TRUE)
   end.date <- max(mydata$date, na.rm = TRUE)
 
-  ## interval in seconds
+  # interval in seconds
   interval <- find.time.interval(mydata$date)
 
-  ## equivalent number of days, used to refine interval for month/year
+  # equivalent number of days, used to refine interval for month/year
   days <- as.numeric(strsplit(interval, split = " ")[[1]][1]) /
     24 /
     3600
 
-  ## find time interval of data
+  # find time interval of data
   if (class(mydata$date)[1] == "Date") {
     interval <- paste(days, "day")
   } else {
-    ## this will be in seconds
+    # this will be in seconds
     interval <- find.time.interval(mydata$date)
   }
 
-  ## better interval, most common interval in a year
+  # better interval, most common interval in a year
   if (days == 31) {
     interval <- "month"
   }
@@ -137,18 +120,18 @@ date.pad <- function(mydata, type = NULL, print.int = FALSE) {
     interval <- "year"
   }
 
-  ## only pad if there are missing data
+  # only pad if there are missing data
   if (length(unique(diff(mydata$date))) != 1L) {
     all.dates <- data.frame(date = seq(start.date, end.date, by = interval))
     mydata <- mydata %>% full_join(all.dates, by = "date")
 
     # add missing types - if type is present
-    if (!is.null(type)) {
+    if (!is.null(type) && type != "default") {
       mydata[type] <- mydata[1, type]
     }
   }
 
-  ## return the same TZ that we started with
+  # return the same TZ that we started with
   attr(mydata$date, "tzone") <- TZ
 
   if (print.int) {
@@ -160,34 +143,8 @@ date.pad <- function(mydata, type = NULL, print.int = FALSE) {
 
   mydata
 }
-#############################################################################################
 
-## unitility function to convert decimal date to POSIXct
-decimalDate <- function(x, date = "date") {
-  thedata <- x
-  x <- x[, date]
-  x.year <- floor(x)
-  ## fraction of the year
-  x.frac <- x - x.year
-  ## number of seconds in each year
-  x.sec.yr <- unclass(ISOdate(x.year + 1, 1, 1, 0, 0, 0)) -
-    unclass(ISOdate(x.year, 1, 1, 0, 0, 0))
-  ## now get the actual time
-  x.actual <- ISOdate(x.year, 1, 1, 0, 0, 0) + x.frac * x.sec.yr
-  x.actual <- as.POSIXct(trunc(x.actual, "hours"), "GMT")
-  thedata$date <- x.actual
-  thedata
-}
-
-convert.date <- function(mydata, format = "%d/%m/%Y %H:%M") {
-  mydata$date <- as.POSIXct(strptime(mydata$date, format = format), "GMT")
-  mydata
-}
-
-
-#############################################################################################
-
-## from Deepayan Sarkar
+# from Deepayan Sarkar
 panel.smooth.spline <-
   function(
     x,
@@ -256,8 +213,7 @@ panel.smooth.spline <-
     }
   }
 
-### panel functions for plots based on lattice ####################################################
-
+# panel functions for plots based on lattice #
 panel.gam <- function(
   x,
   y,
@@ -292,12 +248,12 @@ panel.gam <- function(
   fontface,
   fontfamily
 ) {
-  ## panel function to add a smooth line to a plot
-  ## Uses a GAM (mgcv) to fit smooth
-  ## Optionally can plot 95% confidence intervals and run bootstrap simulations
-  ## to estimate uncertainties. Simple block bootstrap is also available for correlated data
+  # panel function to add a smooth line to a plot
+  # Uses a GAM (mgcv) to fit smooth
+  # Optionally can plot 95% confidence intervals and run bootstrap simulations
+  # to estimate uncertainties. Simple block bootstrap is also available for correlated data
 
-  ## get rid of R check annoyances#
+  # get rid of R check annoyances#
   plot.line <- NULL
 
   thedata <- data.frame(x = x, y = y)
@@ -329,7 +285,7 @@ panel.gam <- function(
         )
         xseq <- seq(xrange[1], xrange[2], length = n)
 
-        ## for uncertainties
+        # for uncertainties
         std <- qnorm(level / 2 + 0.5)
 
         pred <- predict(mod, data.frame(x = xseq), se = TRUE)
@@ -365,7 +321,7 @@ panel.gam <- function(
           pred <- pred$fit
         }
       } else {
-        ## simulations required
+        # simulations required
 
         x <- thedata$x
         y <- thedata$y
@@ -380,7 +336,7 @@ panel.gam <- function(
 
         message("Taking bootstrap samples. Please wait...")
 
-        ## set up bootstrap
+        # set up bootstrap
         block.length <- 1
 
         if (autocor) {
@@ -388,19 +344,19 @@ panel.gam <- function(
         }
         index <- samp.boot.block(sam.size, n.sim, block.length)
 
-        ## predict first
+        # predict first
         if (is.null(k)) {
           mod <- mgcv::gam(y ~ s(x), data = thedata, ...)
         } else {
           mod <- mgcv::gam(y ~ s(x, k = k), data = thedata, ...)
         }
 
-        residuals <- residuals(mod) ## residuals of the model
+        residuals <- residuals(mod) # residuals of the model
 
         pred.input <- predict(mod, thedata)
 
         for (i in 1:n.sim) {
-          ## make new data
+          # make new data
           new.data <- data.frame(
             x = xseq,
             y = pred.input + residuals[index[, i]]
@@ -413,7 +369,7 @@ panel.gam <- function(
           boot.pred[, i] <- as.vector(pred)
         }
 
-        ## calculate percentiles
+        # calculate percentiles
         percentiles <- apply(
           boot.pred,
           1,
@@ -454,44 +410,150 @@ panel.gam <- function(
 }
 
 
-#########################################################################################################
-
-## error in mean from Hmisc
-
-errorInMean <- function(
-  x,
-  mult = qt((1 + conf.int) / 2, n - 1),
-  conf.int = 0.95,
-  na.rm = TRUE
+# version of GAM fitting not for plotting - need to rationalise both...
+fitGam <- function(
+  thedata,
+  x = "date",
+  y = "conc",
+  form = y ~ x,
+  k = k,
+  Args,
+  ...,
+  simulate = FALSE,
+  n.sim = 200,
+  autocor = FALSE,
+  se = TRUE,
+  level = 0.95,
+  n = 100
 ) {
-  if (na.rm) {
-    x <- x[!is.na(x)]
-  }
-  n <- length(x)
-  if (n < 2) {
-    return(c(Mean = mean(x), Lower = NA, Upper = NA))
-  }
-  xbar <- sum(x) / n
-  se <- sqrt(sum((x - xbar)^2) / n / (n - 1))
-  c(
-    Mean = xbar,
-    Lower = xbar - mult * se,
-    Upper = xbar +
-      mult *
-        se
+  # panel function to add a smooth line to a plot
+  # Uses a GAM (mgcv) to fit smooth
+  # Optionally can plot 95% confidence intervals and run bootstrap simulations
+  # to estimate uncertainties. Simple block bootstrap is also available for correlated data
+
+  data.orig <- thedata # return this if all else fails
+
+  id <- which(names(thedata) == x)
+  names(thedata)[id] <- "x"
+  id <- which(names(thedata) == y)
+  names(thedata)[id] <- "y"
+
+  # can only fit numeric, so convert back after fitting
+  class_x <- class(thedata$x)
+
+  thedata$x <- as.numeric(thedata$x)
+
+  tryCatch(
+    {
+      if (!simulate) {
+        if (is.null(k)) {
+          mod <- suppressWarnings(mgcv::gam(
+            y ~ s(x),
+            select = TRUE,
+            data = thedata,
+            ...
+          ))
+        } else {
+          mod <- suppressWarnings(mgcv::gam(
+            y ~ s(x, k = k),
+            select = TRUE,
+            data = thedata,
+            ...
+          ))
+        }
+
+        xseq <- seq(
+          min(thedata$x, na.rm = TRUE),
+          max(thedata$x, na.rm = TRUE),
+          length = n
+        )
+
+        # for uncertainties
+        std <- qnorm(level / 2 + 0.5)
+
+        pred <- predict(mod, data.frame(x = xseq), se = se)
+
+        results <- data.frame(
+          date = xseq,
+          pred = pred$fit,
+          lower = pred$fit - std * pred$se,
+          upper = pred$fit + std * pred$se
+        )
+      } else {
+        # simulations required
+
+        sam.size <- nrow(thedata)
+
+        xseq <- seq(
+          min(thedata$x, na.rm = TRUE),
+          max(thedata$x, na.rm = TRUE),
+          length = n
+        )
+
+        boot.pred <- matrix(nrow = sam.size, ncol = n.sim)
+
+        message("Taking bootstrap samples. Please wait...")
+
+        # set up bootstrap
+        block.length <- 1
+
+        if (autocor) {
+          block.length <- round(sam.size^(1 / 3))
+        }
+        index <- samp.boot.block(sam.size, n.sim, block.length)
+
+        # predict first
+        if (is.null(k)) {
+          mod <- mgcv::gam(y ~ s(x), data = thedata, ...)
+        } else {
+          mod <- mgcv::gam(y ~ s(x, k = k), data = thedata, ...)
+        }
+
+        residuals <- residuals(mod) # residuals of the model
+
+        pred.input <- predict(mod, thedata)
+
+        for (i in 1:n.sim) {
+          # make new data
+          new.data <- data.frame(
+            x = xseq,
+            y = pred.input + residuals[index[, i]]
+          )
+
+          mod <- mgcv::gam(y ~ s(x), data = new.data, ...)
+
+          pred <- predict(mod, new.data)
+
+          boot.pred[, i] <- as.vector(pred)
+        }
+
+        # calculate percentiles
+        percentiles <- apply(
+          boot.pred,
+          1,
+          function(x) quantile(x, probs = c(0.025, 0.975))
+        )
+
+        results <- as.data.frame(cbind(
+          pred = rowMeans(boot.pred),
+          lower = percentiles[1, ],
+          upper = percentiles[2, ]
+        ))
+      }
+
+      # convert class back to original
+      class(results[[x]]) <- class_x
+      return(results)
+    },
+    error = function(x) {
+      data.orig
+    }
   )
 }
 
-###########################################################################################################
 
-## list update function
-## for lattice type object structure and ... handling
-
-## (currently used by)
-## (all openair plots that include colorkey controlled by drawOpenKey)
-
-## listUpdate function
-# [in development]
+# list update function
+# for lattice type object structure and ... handling
 listUpdate <- function(
   a,
   b,
@@ -515,15 +577,9 @@ listUpdate <- function(
   a
 }
 
-#############################################################################################################
-
-## makeOpenKeyLegend v0.1
-
-## common code for making legend list
-## objects for use with drawOpenkey outputs
-
-## uses listUpdate in utilities
-
+# common code for making legend list
+# objects for use with drawOpenkey outputs
+# uses listUpdate in utilities
 makeOpenKeyLegend <- function(key, default.key, fun.name = "function") {
   # handle logicals and lists
   if (is.logical(key)) {
@@ -562,7 +618,7 @@ makeOpenKeyLegend <- function(key, default.key, fun.name = "function") {
   legend
 }
 
-## polygon that can deal with missing data for use in lattice plots with groups
+# polygon that can deal with missing data for use in lattice plots with groups
 poly.na <- function(
   x1,
   y1,
@@ -587,9 +643,9 @@ poly.na <- function(
 }
 
 
-## gives names of lattice strips
+# gives names of lattice strips
 strip.fun <- function(results.grid, type, auto.text) {
-  ## proper names of labelling ###################################################
+  # proper names of labelling #
   pol.name <- sapply(
     levels(factor(results.grid[[type[1]]])),
     function(x) quickText(x, auto.text)
@@ -599,7 +655,7 @@ strip.fun <- function(results.grid, type, auto.text) {
   if (length(type) == 1) {
     strip.left <- FALSE
   } else {
-    ## two conditioning variables
+    # two conditioning variables
 
     pol.name <- sapply(
       levels(factor(results.grid[[type[2]]])),
@@ -609,12 +665,12 @@ strip.fun <- function(results.grid, type, auto.text) {
   }
   if (length(type) == 1 & type[1] == "default") {
     strip <- FALSE
-  } ## remove strip
+  } # remove strip
   list(strip, strip.left, pol.name)
 }
 
 
-## from lattice
+# from lattice
 chooseFace <- function(fontface = NULL, font = 1) {
   if (is.null(fontface)) {
     font
@@ -624,7 +680,7 @@ chooseFace <- function(fontface = NULL, font = 1) {
 }
 
 
-## .smoothScatterCalcDensity() is also in graphics, but not exported.
+# .smoothScatterCalcDensity() is also in graphics, but not exported.
 .smoothScatterCalcDensity <- function(x, nbin, bandwidth, range.x) {
   if (!("KernSmooth" %in% loadedNamespaces())) {
     ns <- try(loadNamespace("KernSmooth"))
@@ -655,7 +711,7 @@ chooseFace <- function(fontface = NULL, font = 1) {
     if (!is.numeric(bandwidth)) stop("'bandwidth' must be numeric")
   }
   bandwidth[bandwidth == 0] <- 1
-  ## create density map
+  # create density map
   if (missing(range.x)) {
     rv <- KernSmooth::bkde2D(x, gridsize = nbin, bandwidth = bandwidth)
   } else {
@@ -671,12 +727,12 @@ chooseFace <- function(fontface = NULL, font = 1) {
 }
 
 
-## simple rounding function from plyr
+# simple rounding function from plyr
 round_any <- function(x, accuracy, f = round) {
   f(x / accuracy) * accuracy
 }
 
-## pretty gap calculator
+# pretty gap calculator
 prettyGap <- function(x, n = 100) {
   return(diff(pretty(x, n))[1])
 }